Player and Computer Hands created, varibles places in main changed, cards now display (+) sign
Player and computer hand is created by taking cards from their respective decks. players' decks are checked if a cards has been taken before by simply checking if it is null or not. Also, I changed the variables places to find them easily if needed. Also, the cards now show if they are (+) signed or not