Changed choices
Player and computer choice is changed. Before, when the turn starts a card was automatically chosen. This was creating a problem when player and computer did not want to ask a card because they are close to 20 but played a card before and couldn't stand. Changed the ask a card when turn begins to making a decision whether ask a card, stand or play a card.