Fixed bug
When there is no card before played but the player choses to play a special card, since there is no card that will affect game was crashing. Now it does not look for a card to change. If there is no card it wont work but player can still play it.